Revert "Merge pull request #83541 from mukhrr/fix/83430"
This reverts commit cfe677f6cc049fc540477476963c1ac7e8641be8, reversing
changes made to 48750b68fc8e57efae4bac75701a4561aea19948.
